requirements.txt
Flask_SocketIO>=5.0.1
flask_restx>=0.2.0
Flask>=1.1.2
Werkzeug>=1.0.1
Flask_Cors>=3.0.10
simplejson>=3.17.2
itsdangerous>=1.1.0
urllib3>=1.26.3
aliyun_python_sdk_core>=2.13.30
dnspython>=2.1.0
pycryptodome>=3.9.9
py-bcrypt>=0.4
reportlab==3.5.67
PyPDF2==1.26.0
sqlalchemy>=1.2.6
Flask-SQLAlchemy>=2.4.1
Flask-Script>=2.0.5
gunicorn>=19.7.0
gevent==20.6.2
requests
pandas==0.25.3
six>=1.15.0
flask_compress==1.4.0
Flask-Migrate==3.0.0
mysqlclient==2.0.3
